requirements.txt
Django==1.11.8
Pillow==4.3.0
boto==2.48.0
dj-database-url==0.4.2
django-bootstrap-breadcrumbs==0.9.0
django-braces==1.12.0
django-forms-bootstrap==3.1.0
django-js-asset==0.1.1
django-model-utils==3.0.0
django-mptt==0.9.0
django-star-ratings==0.5.6
django-storages==1.6.5
gunicorn==19.7.1
olefile==0.44
psycopg2==2.7.3.2
pytz==2017.3
stripe==1.75.1
